Add 'SDATE_GFS' rocoto env var for metp tasks
The connector script between EMC_verif-global and global-workflow
run_verif_global_in_global_workflow.sh (housed in EMC_verif-global)
contains a new feature that checks the maximum forecast hour for
verification is the hour's initialization date is equal to or after
the first GFS cycle initialization. If the date is before, the
maximum forecast hour is adjusted to the maximum hour relative to
the forecast verification date and the first GFS cycle initialization.
This feature was added to cut down on METplus calls being done
on data that don't exist.

Refs: #315